chore(bb): make compile on stock mac clang (#8278)
xcode clang does not support all of c++20 it seems e.g. can't do
Constructor(A,B,C) where A B and C are the members of a struct with only
default constructors. Some common issues that come up like the
uint64_t/size_t split

+++ barretenberg/cpp/src/barretenberg/crypto/merkle_tree/types.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cstdint>
+#include <cstddef>
 namespace bb::crypto::merkle_tree {
-using index_t = uint64_t;
+using index_t = size_t;
 } // namespace bb::crypto::merkle_tree
