+++ barretenberg/cpp/src/barretenberg/vm/avm/tests/slice.test.cpp
@@ -100,7 +100,7 @@ class AvmSliceTests : public ::testing::Test {
         }
 
         // Check that the extra final row is well-formed.
-        EXPECT_THAT(trace.at(static_cast<size_t>(last_row_idx + 1)),
+        EXPECT_THAT(trace.at(static_cast<uint64_t>(last_row_idx + 1)),
                     AllOf(SLICE_ROW_FIELD_EQ(addr, FF(dst_offset) + FF(copy_size)),
                           SLICE_ROW_FIELD_EQ(col_offset, col_offset + copy_size),
                           SLICE_ROW_FIELD_EQ(cnt, 0),
